# openapi.yaml
openapi: 3.0.3
info:
  title: Narada LMS API
  description: |
    Comprehensive API for Narada LMS - A multilingual Learning Management System for Vedic education.
    
    **Architecture:** Modular backend with 6 domain modules:
    - Identity & Access: User management, authentication, authorization
    - Content Publishing: Tracks, chapters, text segmentation
    - Media Pipeline: Audio files, media segments, segment mappings
    - Batch & Cohort: Batches, enrollments, co-instructors, progress evaluation
    - Learning Delivery: Student-facing learning APIs, progress tracking
    - System Admin: Audit logs, system settings, monitoring
    
    **Multilingual Support:** Telugu (te), Hindi/Devanagari (hi), English/IAST (en)
    
    **Authentication:** JWT Bearer tokens. Obtain a token via ``POST /api/auth/login``. Protected endpoints require the ``Authorization: Bearer <token>`` header.
    
    **Authorization:** Role-based access control (student, instructor, content_manager, admin)
  version: 2.0.0
  contact:
    name: NaradaLMS Support
    email: support@narada-lms.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/media-segments/bulk:
    post:
      tags: [Media Segments]
      summary: Create media segments in bulk
      description: Create multiple media segments at once - Content managers/Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audioFileId, segments]
              properties:
                audioFileId:
                  type: integer
                segments:
                  type: array
                  items:
                    type: object
                    required: [startMs, endMs]
                    properties:
                      startMs:
                        type: integer
                        minimum: 0
                        description: Start timestamp in milliseconds
                      endMs:
                        type: integer
                        minimum: 0
                        description: End timestamp in milliseconds
      responses:
        '201':
          description: Media segments created successfully



  # =============================================================================
  # BATCH & COHORT (batch-cohort module)
  # =============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /api/auth/login; send as Authorization Bearer <token>
...